updating controllers for mongoose/new auth

+++ public/project/client/services/recommend.service.client.js
@@ -14,6 +14,10 @@
         return api;
 
         function createRecommend(userId, reco) {
+            if(!userId) {
+                console.log("Must specify a userId to create reco.")
+                return;
+            }
             return $http({
                 method: 'POST',
                 url: '/api/project/user/'+userId+'/recommend',
@@ -22,11 +26,19 @@
         }
 
         function deleteRecommendation(userId, recoId) {
+            if(!userId) {
+                console.log("Must specify a userId to delete reco.")
+                return;
+            }
             return $http.delete('/api/project/user/'+userId+'/recommend'+recoId);
         }
 
 
         function getRecommendsForUserId(userId) {
+            if(!userId) {
+                console.log("Must specify a userId to get recos.")
+                return;
+            }
             return $http.get('/api/project/user/'+userId+'/recommend');
         }
     }
